+++ src/main/java/io/inscopemetrics/kairosdb/aggregators/DelegatingAggregator.java
@@ -29,7 +29,7 @@ import javax.inject.Provider;
 /**
  * Serves as a base for an aggregator that will delegate and intelligently dispatch aggregation to appropriate classes.
  *
- * @author Brandon Arp (brandon dot arp at smartsheet dot com)
+ * @author Brandon Arp (brandon dot arp at inscopemetrics dot io)
  */
 public class DelegatingAggregator implements Aggregator {
     private static final Logger LOGGER = LoggerFactory.getLogger(DelegatingAggregator.class);