+++ src/main/java/io/inscopemetrics/kairosdb/aggregators/DelegatingPercentileAggregator.java
@@ -29,7 +29,7 @@ import javax.inject.Named;
 /**
  * Percentile aggregator that delegates to the built-in and the histogram aggregators.
  *
- * @author Brandon Arp (brandon dot arp at smartsheet dot com)
+ * @author Brandon Arp (brandon dot arp at inscopemetrics dot io)
  */
 @FeatureComponent(
         name = "percentile",